consolidate dashboard component
