+++ src/components/SummaryCards.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import SummaryCard from 'components/SummaryCard';
 import { Istats } from 'utils/types';
 
@@ -8,10 +7,10 @@ interface SummaryCardsProps {
 
 const SummaryCards = ({ stats } : SummaryCardsProps) => {
   const cardData = [
-    { title: 'Top Grade', value: stats?.highestGrade, className: 'summary-card-green' },
-    { title: 'Most Passed', value: stats?.mostPassedSubject, className: 'summary-card-green' },
-    { title: 'Lowest Grade', value: stats?.lowestGrade, className: 'summary-card-red' },
-    { title: 'Most Failed', value: stats?.mostFailedSubject, className: 'summary-card-red' }
+    { title: 'Top Grade', value: stats?.highestGrade|| '--', className: 'summary-card-green' },
+    { title: 'Most Passed', value: stats?.mostPassedSubject|| '--', className: 'summary-card-green' },
+    { title: 'Lowest Grade', value: stats?.lowestGrade|| '--', className: 'summary-card-red' },
+    { title: 'Most Failed', value: stats?.mostFailedSubject|| '--', className: 'summary-card-red' }
   ];
 
   return (
