+++ src/components/Table/Table.tsx
@@ -1,18 +1,47 @@
 import React from 'react';
+import { useNavigate } from 'react-router-dom';
+
 import { IResultData } from 'utils/types';
-import TableRow from './TableRow';
-import TableHeader from './TableHeader';
+import TableRow from 'components/Table/TableRow';
+import TableHeader from 'components/Table/TableHeader';
 
 interface TableProps {
   students: IResultData[];
-  handleActionClick: (student: IResultData) => void;
   dropdownVisible: boolean;
+  setDropdownVisible: React.Dispatch<React.SetStateAction<boolean>>;
   selectedStudent: IResultData | null;
+  setSelectedStudent: React.Dispatch<React.SetStateAction<IResultData | null>>;
   dropdownRef: React.RefObject<HTMLDivElement>;
-  handleEdit: () => void;
-  handleDelete: () => void;
+  setShowModal: React.Dispatch<React.SetStateAction<boolean>>;
 }
-function Table ({ students, handleActionClick, dropdownVisible, selectedStudent, dropdownRef, handleEdit, handleDelete }: TableProps) {
+
+const Table = ({ students, dropdownVisible, setDropdownVisible, selectedStudent, setSelectedStudent, dropdownRef, setShowModal }: TableProps) => {
+  const navigate = useNavigate();
+
+  const handleActionClick = (student: IResultData) => {
+    setSelectedStudent(student);
+    setDropdownVisible(!dropdownVisible);
+    if (dropdownRef.current) {
+      const dropdownRect = dropdownRef.current.getBoundingClientRect();
+      if (dropdownRect.bottom > window.innerHeight) {
+        dropdownRef.current.style.top = `-${dropdownRect.height}px`;
+      } else {
+        dropdownRef.current.style.top = `100%`;
+      }
+    }
+  };
+
+  const handleEdit = () => {
+    if (selectedStudent) {
+      navigate('/result/add', { state: { student: selectedStudent, isUpdate: true } });
+    }
+    setDropdownVisible(false);
+  };
+
+  const handleDelete = () => {
+    setShowModal(true);
+  };
+
   return (
     <div className="table-section">
       <TableHeader />
@@ -30,6 +59,6 @@ function Table ({ students, handleActionClick, dropdownVisible, selectedStudent,
       ))}
     </div>
   );
-};
+}
 
 export default Table;
