+++ src/components/Table/TableRow.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import ActionIcon from 'components/Icons/ActionIcon';
 import ActionMenu from 'components/ActionMenu';
 import { getGradeClassName } from 'utils/grades';
